move cross test utilities to be shared between sdkv2 and pf
