+++ pkg/internal/tests/cross-tests/diff_check.go
@@ -12,21 +12,15 @@
 // See the License for the specific language governing permissions and
 
 // Compares the effect of transitioning between two randomly sampled resource configurations.
-//
-//nolint:lll
 package crosstests
 
 import (
 	"encoding/json"
-	"fmt"
 	"os"
 	"path/filepath"
 
 	"github.com/hashicorp/terraform-plugin-go/tftypes"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/pulumi/pulumi/sdk/v3/go/auto"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/apitype"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	crosstestsimpl "github.com/pulumi/pulumi-terraform-bridge/v3/pkg/internal/tests/cross-tests/impl"
@@ -51,17 +45,7 @@ type diffTestCase struct {
 	DeleteBeforeReplace bool
 }
 
-type pulumiDiffResp struct {
-	DetailedDiff        map[string]interface{} `json:"detailedDiff"`
-	DeleteBeforeReplace bool                   `json:"deleteBeforeReplace"`
-}
-
-type diffResult struct {
-	TFDiff     tfChange
-	PulumiDiff pulumiDiffResp
-}
-
-func runDiffCheck(t T, tc diffTestCase) diffResult {
+func runDiffCheck(t T, tc diffTestCase) crosstestsimpl.DiffResult {
 	t.Helper()
 	tfwd := t.TempDir()
 
@@ -97,76 +81,19 @@ func runDiffCheck(t T, tc diffTestCase) diffResult {
 	require.NoErrorf(t, err, "writing Pulumi.yaml")
 	x := pt.Up(t)
 
-	changes := tfd.parseChangesFromTFPlan(*tfDiffPlan)
+	changes := tfd.driver.ParseChangesFromTFPlan(tfDiffPlan)
 
-	diffResponse := pulumiDiffResp{}
+	diffResponse := crosstestsimpl.PulumiDiffResp{}
 	for _, entry := range pt.GrpcLog(t).Entries {
 		if entry.Method == "/pulumirpc.ResourceProvider/Diff" {
 			err := json.Unmarshal(entry.Response, &diffResponse)
 			require.NoError(t, err)
 		}
 	}
-	tc.verifyBasicDiffAgreement(t, changes.Actions, x.Summary, diffResponse)
+	crosstestsimpl.VerifyBasicDiffAgreement(t, changes.Actions, x.Summary, diffResponse)
 
-	return diffResult{
+	return crosstestsimpl.DiffResult{
 		TFDiff:     changes,
 		PulumiDiff: diffResponse,
-	}
-}
-
-func (tc *diffTestCase) verifyBasicDiffAgreement(t T, tfActions []string, us auto.UpdateSummary, diffResponse pulumiDiffResp) {
-	t.Helper()
-	t.Logf("UpdateSummary.ResourceChanges: %#v", us.ResourceChanges)
-	// Action list from https://github.com/opentofu/opentofu/blob/main/internal/plans/action.go#L11
-	if len(tfActions) == 0 {
-		require.FailNow(t, "No TF actions found")
-	}
-	if len(tfActions) == 1 {
-		switch tfActions[0] {
-		case "no-op":
-			require.NotNilf(t, us.ResourceChanges, "UpdateSummary.ResourceChanges should not be nil")
-			rc := *us.ResourceChanges
-			assert.Equalf(t, 2, rc[string(apitype.OpSame)], "expected the test resource and stack to stay the same")
-			assert.Equalf(t, 1, len(rc), "expected one entry in UpdateSummary.ResourceChanges")
-		case "create":
-			require.NotNilf(t, us.ResourceChanges, "UpdateSummary.ResourceChanges should not be nil")
-			rc := *us.ResourceChanges
-			assert.Equalf(t, 1, rc[string(apitype.OpSame)], "expected the stack to stay the same")
-			assert.Equalf(t, 1, rc[string(apitype.OpCreate)], "expected the test resource to get a create plan")
-		case "read":
-			require.FailNow(t, "Unexpected TF action: read")
-		case "update":
-			require.NotNilf(t, us.ResourceChanges, "UpdateSummary.ResourceChanges should not be nil")
-			rc := *us.ResourceChanges
-			assert.Equalf(t, 1, rc[string(apitype.OpSame)], "expected one resource to stay the same - the stack")
-			assert.Equalf(t, 1, rc[string(apitype.Update)], "expected the test resource to get an update plan")
-			assert.Equalf(t, 2, len(rc), "expected two entries in UpdateSummary.ResourceChanges")
-		case "delete":
-			require.NotNilf(t, us.ResourceChanges, "UpdateSummary.ResourceChanges should not be nil")
-			rc := *us.ResourceChanges
-			assert.Equalf(t, 1, rc[string(apitype.OpSame)], "expected the stack to stay the same")
-			assert.Equalf(t, 1, rc[string(apitype.OpDelete)], "expected the test resource to get a delete plan")
-		default:
-			panic("TODO: do not understand this TF action yet: " + tfActions[0])
-		}
-	} else if len(tfActions) == 2 {
-		if tfActions[0] == "create" && tfActions[1] == "delete" {
-			require.NotNilf(t, us.ResourceChanges, "UpdateSummary.ResourceChanges should not be nil")
-			rc := *us.ResourceChanges
-			assert.Equalf(t, 1, rc[string(apitype.OpSame)], "expected the stack to stay the same")
-			assert.Equalf(t, 1, rc[string(apitype.OpReplace)], "expected the test resource to get a replace plan")
-			assert.Equalf(t, false, diffResponse.DeleteBeforeReplace, "expected deleteBeforeReplace to be true")
-		} else if tfActions[0] == "delete" && tfActions[1] == "create" {
-			require.NotNilf(t, us.ResourceChanges, "UpdateSummary.ResourceChanges should not be nil")
-			rc := *us.ResourceChanges
-			t.Logf("UpdateSummary.ResourceChanges: %#v", rc)
-			assert.Equalf(t, 1, rc[string(apitype.OpSame)], "expected the stack to stay the same")
-			assert.Equalf(t, 1, rc[string(apitype.OpReplace)], "expected the test resource to get a replace plan")
-			assert.Equalf(t, true, diffResponse.DeleteBeforeReplace, "expected deleteBeforeReplace to be true")
-		} else {
-			panic("TODO: do not understand this TF action yet: " + fmt.Sprint(tfActions))
-		}
-	} else {
-		panic("TODO: do not understand this TF action yet: " + fmt.Sprint(tfActions))
 	}
 }
