+++ pkg/internal/tests/cross-tests/tf_driver.go
@@ -17,11 +17,9 @@ package crosstests
 
 import (
 	"bytes"
-	"encoding/json"
 
 	"github.com/hashicorp/terraform-plugin-go/tftypes"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/util/contract"
 	"github.com/stretchr/testify/require"
 	"github.com/zclconf/go-cty/cty"
 
@@ -31,7 +29,7 @@ import (
 )
 
 type TfResDriver struct {
-	driver *tfcheck.TfDriver
+	driver *tfcheck.TFDriver
 	res    *schema.Resource
 }
 
@@ -84,7 +82,7 @@ func (d *TfResDriver) writePlanApply(
 	resourceType, resourceName string,
 	config cty.Value,
 	lifecycle lifecycleArgs,
-) *tfcheck.TfPlan {
+) *tfcheck.TFPlan {
 	if !config.IsNull() {
 		d.write(t, resourceSchema, resourceType, resourceName, config, lifecycle)
 	} else {
@@ -121,32 +119,6 @@ func (d *TfResDriver) write(
 	require.NoError(t, err)
 	t.Logf("HCL: \n%s\n", buf.String())
 	d.driver.Write(t, buf.String())
-}
-
-type tfChange struct {
-	Actions []string       `json:"actions"`
-	Before  map[string]any `json:"before"`
-	After   map[string]any `json:"after"`
-}
-
-// Still discovering the structure of JSON-serialized TF plans. The information required from these is, primarily, is
-// whether the resource is staying unchanged, being updated or replaced. Secondarily, would be also great to know
-// detailed paths of properties causing the change, though that is more difficult to cross-compare with Pulumi.
-//
-// For now this is code is similar to `jq .resource_changes[0].change.actions[0] plan.json`.
-func (*TfResDriver) parseChangesFromTFPlan(plan tfcheck.TfPlan) tfChange {
-	type p struct {
-		ResourceChanges []struct {
-			Change tfChange `json:"change"`
-		} `json:"resource_changes"`
-	}
-	jb, err := json.Marshal(plan.RawPlan)
-	contract.AssertNoErrorf(err, "failed to marshal terraform plan")
-	var pp p
-	err = json.Unmarshal(jb, &pp)
-	contract.AssertNoErrorf(err, "failed to unmarshal terraform plan")
-	contract.Assertf(len(pp.ResourceChanges) == 1, "expected exactly one resource change")
-	return pp.ResourceChanges[0].Change
 }
 
 func providerHCLProgram(t T, typ string, provider *schema.Provider, config cty.Value) string {
