+++ pkg/pf/tests/internal/cross-tests/util.go
@@ -21,11 +21,16 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/pf/tests/internal/providerbuilder"
+	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/pf/tfbridge"
+	tfbridge0 "github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge"
+	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge/info"
+	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge/tokens"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/diag"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/resource"
 )
 
-func propageteSkip(parent, child *testing.T) {
+func propagateSkip(parent, child *testing.T) {
 	if child.Skipped() {
 		parent.Skipf("skipping due to skipped child test")
 	}
@@ -53,4 +58,19 @@ func skipUnlessLinux(t *testing.T) {
 	if ci, ok := os.LookupEnv("CI"); ok && ci == "true" && !strings.Contains(strings.ToLower(runtime.GOOS), "linux") {
 		t.Skip("Skipping on non-Linux platforms as our CI does not yet install Terraform CLI required for these tests")
 	}
+}
+
+func bridgedProvider(prov *providerbuilder.Provider) info.Provider {
+	shimProvider := tfbridge.ShimProvider(prov)
+
+	provider := tfbridge0.ProviderInfo{
+		P:            shimProvider,
+		Name:         prov.TypeName,
+		Version:      prov.Version,
+		MetadataInfo: &tfbridge0.MetadataInfo{},
+	}
+
+	provider.MustComputeTokens(tokens.SingleModule(prov.TypeName, "index", tokens.MakeStandard(prov.TypeName)))
+
+	return provider
 }
